requirements.txt
gpxpy==1.5.0
python-dateutil==2.8.2
six==1.16.0
websockets==10.1
